SYNC-1569 Add 'pending wipe' and change status to wiped when client attempts to sync

+++ webs/sync-web/docroot/WEB-INF/src/com/liferay/sync/service/impl/SyncDeviceLocalServiceImpl.java
@@ -37,7 +37,7 @@ public class SyncDeviceLocalServiceImpl extends SyncDeviceLocalServiceBaseImpl {
 
 	@Override
 	public SyncDevice addSyncDevice(
-			long userId, String type, int buildNumber, int featureSet)
+			long userId, String type, long buildNumber, int featureSet)
 		throws PortalException, SystemException {
 
 		User user = userPersistence.findByPrimaryKey(userId);
@@ -87,7 +87,7 @@ public class SyncDeviceLocalServiceImpl extends SyncDeviceLocalServiceBaseImpl {
 
 	@Override
 	public SyncDevice updateSyncDevice(
-			long syncDeviceId, String type, int buildNumber, int featureSet,
+			long syncDeviceId, String type, long buildNumber, int featureSet,
 			int status)
 		throws PortalException, SystemException {
