+++ webs/sync-web/docroot/WEB-INF/src/com/liferay/sync/service/impl/SyncDeviceServiceImpl.java
@@ -32,7 +32,7 @@ public class SyncDeviceServiceImpl extends SyncDeviceServiceBaseImpl {
 
 	@Override
 	public SyncDevice registerSyncDevice(
-			String type, int buildNumber, int featureSet, String uuid)
+			String type, long buildNumber, int featureSet, String uuid)
 		throws PortalException, SystemException {
 
 		User user = getUser();
